fix errors

+++ Ccode.py
@@ -32,10 +32,11 @@ try:
     file.close()
 except:
     print("\n\n\n\ngiven file missing, aborting\n\n\n\n")
-    os.system("rm -r node_modules")
-    os.system("rm package.json")
-    os.system("rm package-lock.json")
-    os.system("rm main.js")
+    if not node:
+        os.system("rm -r node_modules")
+        os.system("rm package.json")
+        os.system("rm package-lock.json")
+        os.system("rm main.js")
 
 
 #unknown command message
